# swagger/swagger.yaml
definitions:
  handler.AvatarReq:
    properties:
      avatar_url:
        description: 头像URL
        type: string
    type: object
  handler.LoginReq:
    properties:
      account:
        description: 账号（邮箱/手机号）
        type: string
      password:
        description: 密码
        type: string
    type: object
  handler.PasswordReq:
    properties:
      new_password:
        description: 新密码
        type: string
      old_password:
        description: 旧密码
        type: string
    type: object
  handler.RegisterReq:
    properties:
      email:
        description: 邮箱（可选）
        type: string
      password:
        description: 密码（必填）
        type: string
      phone:
        description: 手机号（可选）
        type: string
      username:
        description: 用户名（可选）
        type: string
    type: object
  handler.Resp:
    properties:
      code:
        description: 错误码，0表示成功
        type: integer
      data:
        description: 业务数据
      message:
        description: 状态描述
        type: string
    type: object
  handler.StatusReq:
    properties:
      status:
        description: 状态（0禁用/1正常）
        type: integer
    type: object
info:
  contact: {}
  description: Manjing AI backend API.
  title: Manjing AI API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 用户登录
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      security:
      - BearerAuth: []
      summary: 退出登录
      tags:
      - Auth
  /api/v1/auth/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: 修改密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - Auth
  /api/v1/auth/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      summary: 用户注册
      tags:
      - Auth
  /api/v1/users/{id}/avatar:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 头像
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AvatarReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      security:
      - BearerAuth: []
      summary: 更新用户头像
      tags:
      - User
  /api/v1/users/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.StatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Resp'
      security:
      - BearerAuth: []
      summary: 更新用户状态
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"